type clean up and add reference native script support
