+++ packages/sidan-csl-rs/src/model/tx_builder_types/datum.rs
@@ -1,0 +1,17 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub enum Datum {
+    Inline(String),
+    Hash(String),
+}
+
+impl Datum {
+    pub fn get_inner(&self) -> &str {
+        match self {
+            Datum::Inline(s) => s,
+            Datum::Hash(s) => s,
+        }
+    }
+}
