+++ packages/sidan-csl-rs/src/model/tx_builder_types/metadata.rs
@@ -1,0 +1,8 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct Metadata {
+    pub tag: String,
+    pub metadata: String,
+}
