+++ packages/sidan-csl-rs/src/model/tx_builder_types/redeemer.rs
@@ -1,0 +1,15 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct Redeemer {
+    pub data: String,
+    pub ex_units: Budget,
+}
+
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct Budget {
+    pub mem: u64,
+    pub steps: u64,
+}
