+++ packages/sidan-csl-rs/src/model/tx_builder_types/tx_in.rs
@@ -1,0 +1,66 @@
+use serde::{Deserialize, Serialize};
+
+use crate::model::{Asset, Redeemer};
+
+use super::{DatumSource, InlineSimpleScriptSource, ProvidedSimpleScriptSource, ScriptSource};
+
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub enum TxIn {
+    PubKeyTxIn(PubKeyTxIn),
+    SimpleScriptTxIn(SimpleScriptTxIn),
+    ScriptTxIn(ScriptTxIn),
+}
+
+
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct RefTxIn {
+    pub tx_hash: String,
+    pub tx_index: u32,
+}
+
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct PubKeyTxIn {
+    pub tx_in: TxInParameter,
+}
+
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct SimpleScriptTxIn {
+    pub tx_in: TxInParameter,
+    pub simple_script_tx_in: SimpleScriptTxInParameter,
+}
+
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub enum SimpleScriptTxInParameter {
+    ProvidedSimpleScriptSource(ProvidedSimpleScriptSource),
+    InlineSimpleScriptSource(InlineSimpleScriptSource),
+}
+
+
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct ScriptTxIn {
+    pub tx_in: TxInParameter,
+    pub script_tx_in: ScriptTxInParameter,
+}
+
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct TxInParameter {
+    pub tx_hash: String,
+    pub tx_index: u32,
+    pub amount: Option<Vec<Asset>>,
+    pub address: Option<String>,
+}
+
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct ScriptTxInParameter {
+    pub script_source: Option<ScriptSource>,
+    pub datum_source: Option<DatumSource>,
+    pub redeemer: Option<Redeemer>,
+}
