+++ packages/sidan-csl-rs/src/model/tx_builder_types/validity_range.rs
@@ -1,0 +1,8 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct ValidityRange {
+    pub invalid_before: Option<u64>,
+    pub invalid_hereafter: Option<u64>,
+}
