+++ packages/sidan-csl-rs/src/model/tx_builder_types/withdrawal.rs
@@ -1,0 +1,26 @@
+use serde::{Deserialize, Serialize};
+
+use super::{Redeemer, ScriptSource};
+
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub enum Withdrawal {
+    PubKeyWithdrawal(PubKeyWithdrawal),
+    PlutusScriptWithdrawal(PlutusScriptWithdrawal),
+}
+
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct PubKeyWithdrawal {
+    pub address: String,
+    pub coin: u64,
+}
+
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct PlutusScriptWithdrawal {
+    pub address: String,
+    pub coin: u64,
+    pub script_source: Option<ScriptSource>,
+    pub redeemer: Option<Redeemer>,
+}
